refactor: abstract hooks from transfer panel
